Remove duplicated memory_resource_tests (#1451)
During the initial introduction of `async_resource_ref` we duplicated all tests that take a `device_memory_resource*`.

Now that we have already some experience with running it in production and are moving more of the interfaces to `async resource_ref` remove those duplicated tests.

closes #1450

Authors:
  - Michael Schellenberger Costa (https://github.com/miscco)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rmm/pull/1451
