+++ tests/mr/device/thrust_allocator_tests.cu
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "mr_test.hpp"
+#include "mr_ref_test.hpp"
 
 #include <rmm/cuda_stream_view.hpp>
 #include <rmm/device_vector.hpp>
@@ -32,7 +32,7 @@ template class rmm::mr::thrust_allocator<int>;
 namespace rmm::test {
 namespace {
 
-struct allocator_test : public mr_test {};
+struct allocator_test : public mr_ref_test {};
 
 TEST_P(allocator_test, first)
 {
